pyproject.toml
[tool.poetry]
name = "fasttelemail"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
redis = "^5.2.1"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
fastapi-mail = "^1.4.2"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
flower = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
